Extract drag&drop callback to a separate class
